--- feature/main/src/main/java/kr/co/main/calendar/ui/calendarScreen/addScheduleScreen/schedule_input/ScheduleGeneralInput.kt
@@ -1,4 +1,0 @@
-package kr.co.main.calendar.ui.calendarScreen.addScheduleScreen.schedule_input
-
-class ScheduleGeneralInput {
-}
