+++ feature/main/src/main/java/kr/co/main/calendar/ui/screen/addDiaryScreen/diary_input/DiaryDateInput.kt
@@ -1,4 +1,4 @@
-package kr.co.main.calendar.ui.calendarScreen.addDiaryScreen.diary_input
+package kr.co.main.calendar.ui.screen.addDiaryScreen.diary_input
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Surface
