+++ feature/main/src/main/java/kr/co/main/calendar/ui/screen/addDiaryScreen/diary_input/DiaryMemoInput.kt
@@ -1,18 +1,23 @@
-package kr.co.main.calendar.ui.calendarScreen.addScheduleScreen.schedule_input
+package kr.co.main.calendar.ui.screen.addDiaryScreen.diary_input
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import kr.co.main.R
+import kr.co.main.calendar.ui.common.CalendarContainerTextField
 import kr.co.ui.theme.colors
 import kr.co.ui.theme.typo
 
 @Composable
-internal fun ScheduleMemoInput(
+internal fun DiaryMemoInput(
     memo: String,
     onMemoInput: (String) -> Unit,
     modifier: Modifier = Modifier
@@ -38,18 +43,18 @@ private fun MemoTextField(
     onMemoInput: (String) -> Unit,
     modifier: Modifier = Modifier
 ) {
-//    CalendarContainerTextField(
-//        modifier = modifier,
-//        value = memo,
-//        onValueChange = onMemoInput,
-//        placeHolder = {
-//            Text(
-//                text = stringResource(id = R.string.feature_main_calendar_add_schedule_input_memo),
-//                style = MaterialTheme.typo.bodyM,
-//                color = MaterialTheme.colors.text1
-//            )
-//        }
-//    )
+    CalendarContainerTextField(
+        modifier = modifier,
+        value = memo,
+        onValueChange = onMemoInput,
+        placeHolder = {
+            Text(
+                text = stringResource(id = R.string.feature_main_calendar_add_diary_input_memo),
+                style = MaterialTheme.typo.bodyM,
+                color = MaterialTheme.colors.text1
+            )
+        }
+    )
 }
 
 @Composable
@@ -63,4 +68,15 @@ private fun MemoTextCount(
         style = MaterialTheme.typo.labelM,
         color = if (memo.length > 1000) MaterialTheme.colors.red1 else MaterialTheme.colors.text2
     )
+}
+
+@Preview
+@Composable
+private fun DiaryMemoInputPreview() {
+    Surface {
+        DiaryMemoInput(
+            memo = "",
+            onMemoInput = { _ -> }
+        )
+    }
 }
