+++ feature/main/src/main/java/kr/co/main/calendar/ui/screen/addScheduleScreen/schedule_input/ScheduleGeneralInput.kt
@@ -1,0 +1,4 @@
+package kr.co.main.calendar.ui.screen.addScheduleScreen.schedule_input
+
+class ScheduleGeneralInput {
+}
