+++ feature/main/src/main/java/kr/co/main/calendar/ui/screen/calendarScreen/diaryTab/DiaryTab.kt
@@ -1,4 +1,4 @@
-package kr.co.main.calendar.ui.calendarScreen.calendarScreen.diaryTab
+package kr.co.main.calendar.ui.screen.calendarScreen.diaryTab
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
